Refactor file and component names
