+++ ui/pages/pages.scss
@@ -29,7 +29,7 @@
 @import 'send/send';
 @import 'settings/index';
 @import 'snaps/snaps-list-tab/index';
-@import 'snaps/view-snap/index';
+@import 'snaps/snap-view/index';
 @import 'swaps/index';
 @import 'token-allowance/index';
 @import 'token-details/index';
