+++ ui/pages/snaps/snap-view/constants.ts
