+++ ui/pages/snaps/snap-view/index.js
@@ -1,0 +1,1 @@
+export { default } from './snap-view';
