+++ ui/pages/snaps/snap-view/index.scss
