+++ ui/pages/snaps/snaps-list-tab/index.scss
@@ -68,7 +68,7 @@
     }
   }
 
-  &__list-tab {
+  &__list {
     &__wrapper {
       @include screen-md-min {
         margin-top: 24px;
