+++ ui/pages/snaps/snaps-list-tab/snap-list.stories.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { Provider } from 'react-redux';
 import configureStore from '../../../store/store';
 import testData from '../../../../.storybook/test-data';
-import SnapListTab from './snap-list-tab';
+import SnapList from './snap-list';
 
 // Using Test Data For Redux
 const store = configureStore(testData);
@@ -27,7 +27,7 @@ export const DefaultStory = (args) => {
 
   return (
     <div>
-      <SnapListTab
+      <SnapList
         {...args}
         snaps={state.metamask.snaps}
         viewingSnap={viewingSnap}
