--- ui/pages/snaps/view-snap/index.js
@@ -1,1 +1,0 @@
-export { default } from './view-snap';
